switch to longer, int type on long

+++ src/Main.java
@@ -1,9 +1,9 @@
 public class Main {
 
     public static void main(String[] args) {
-        int[] sales = new int[] {100, 200, 300, 400, 500, 600};
+        long[] sales = new long[] {100, 200, 300, 400, 500, 600};
         SalesManager salesManager = new SalesManager(sales);
-        int max = salesManager.max();
+        long max = salesManager.max();
         System.out.println(max);
     }
 }
